float is req to cal mean

+++ model.py
@@ -46,8 +46,8 @@ class LayerNormalization(nn.Module):
         self.beta = nn.Parameter(torch.zeros(1)) # Added
     
     def forward(self , x):
-        mean = x.mean(dim = -1 , keepdim = True)
-        std = x.std(dim = -1 , keepdim = True)
+        mean = x.float().mean(dim = -1 , keepdim = True)
+        std = x.float().std(dim = -1 , keepdim = True)
         return self.alpha*(x-mean) / (std+self.eps)  + self.beta
 
 class FeedForwardBlock(nn.Module):
